Increase tile state limit to Integer.MAX_VALUE, also break existing worlds :>

+++ src/main/java/de/ellpeck/rockbottom/api/world/DynamicRegistryInfo.java
@@ -43,7 +43,7 @@ public class DynamicRegistryInfo{
     }
 
     public void fromBuffer(ByteBuf buf){
-        this.tileRegInfo = new NameToIndexInfo("tile_reg_client_world", null, Short.MAX_VALUE);
+        this.tileRegInfo = new NameToIndexInfo("tile_reg_client_world", null, Integer.MAX_VALUE);
         this.tileRegInfo.fromBuffer(buf);
 
         this.biomeRegInfo = new NameToIndexInfo("biome_reg_client_world", null, Short.MAX_VALUE);
